refactor: move exercise data to components
